requirements.txt
streamlit==1.34.0
opencv-python-headless==4.9.0.80
numpy==1.24.4
pickle5==0.0.11
tensorflow==2.13.0
keras==2.13.1
keras-facenet==0.3.2
mtcnn==0.1.0
scipy==1.10.1
pillow==9.5.0
requests==2.31.0
